+++ src/main/java/com/excilys/formation/computerdatabase/service/Service.java
@@ -8,42 +8,52 @@ import com.excilys.formation.computerdatabase.exception.NotImplementedMethodExce
 /**
  * @author GUIDS
  *
- * @param <T> : the object of the service
+ * @param <T>
+ *          : the object of the service
  */
 public interface Service<T> {
 
-  /** Create a T object in the database. 
+  /**
+   * Create a T object in the database.
    * 
-   * @param object : the object to create
+   * @param object
+   *          : the object to create
    * @return : the T object created
    */
   public default T insert(T object) throws NotImplementedMethodException {
     throw new NotImplementedMethodException("Default implementation of create");
   }
 
-  /** Update a T object in the database.
+  /**
+   * Update a T object in the database.
    * 
-   * @param object : the object to update
+   * @param object
+   *          : the object to update
    * @return : the T object updated
    */
   public default T update(T object) throws NotImplementedMethodException {
     throw new NotImplementedMethodException("Default implementation of update");
   }
 
-  /** Delete a T object in the database.
+  /**
+   * Delete a T object in the database.
    * 
-   * @param objectId : the id of the T object to delete
+   * @param objectId
+   *          : the id of the T object to delete
    */
-  public default void delete(int objecId) throws NotImplementedMethodException {
+  public default void delete(long objecId) throws NotImplementedMethodException {
     throw new NotImplementedMethodException("Default implementation of delete");
   }
 
-  /** Get the list of all the T objects in the database.
+  /**
+   * Get the list of all the T objects in the database.
    * 
-   * @param nbElements : the number of T objects in the database
-   * @param offset : the offset to display
+   * @param nbElements
+   *          : the number of T objects in the database
+   * @param offset
+   *          : the offset to display
    * @return : the list of all the T objects in the database
    */
-  public List<T> list(int nbElements, int offset); 
+  public List<T> list(int nbElements, int offset);
 
 }
